// src/modules/autos/domain/auto.enum.ts
export enum EstadoAuto {
  DISPONIBLE = 'DISPONIBLE',
  RESERVADO = 'RESERVADO',
  VENDIDO = 'VENDIDO',
  POR_INGRESAR = 'POR_INGRESAR',
}

export enum Transmision {
  MANUAL = 'MANUAL',
  AUTOMATICA = 'AUTOMATICA',
}

export enum Color {
  BLANCO = 'BLANCO',
  NEGRO = 'NEGRO',
  GRIS = 'GRIS',
  ROJO = 'ROJO',
  AZUL = 'AZUL',
  VERDE = 'VERDE',
  AMARILLO = 'AMARILLO',
  OTRO = 'OTRO',
}